include root span to the search
